Fix sending a mail on missing required role
Previously, we would register an error in the login context
(which is a login context because we are in the login flow)
when the required role check failed. This would then lead to
an email being sent when the email event listener is active,
as that catches login errors by default.

Since this is not really useful for the users and might confuse
them more than it helps, we decided to remove that feature.

Sadly, we loose the ability to track those failed logins in the
console though, as that was triggered by that login error event
as well.

Additionally, we now use failureChallenge instead of forceChallenge
as that one is similar to the forceChallenge, but also counts
the amount of failed events for brute force protection.